% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{add_biMAP_sce}
\alias{add_biMAP_sce}
\title{Add cacomp obj results to SingleCellExperiment object}
\usage{
add_biMAP_sce(sce, umap_coords, biMAP_meta_name = "biMAP")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{umap_coords}{data.frame with coordinates of genes and cells}

\item{biMAP_meta_name}{name not listed in colData(sce), rowData(sce), or metadata(sce)}
}
\description{
Add cacomp obj results to SingleCellExperiment object
}
