% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{add_caobj_sce}
\alias{add_caobj_sce}
\title{Add cacomp obj results to SingleCellExperiment object}
\usage{
add_caobj_sce(sce, caobj, cacomp_meta_name = "caobj")
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{caobj}{caclust::caclust object}

\item{cacomp_meta_name}{column name not listed in colData(sce), rowData(sce), or metadata(sce)}
}
\description{
Add cacomp obj results to SingleCellExperiment object
}
