% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_zero_dim}
\alias{add_zero_dim}
\title{Add an extra 0 dimension to vector}
\usage{
add_zero_dim(vectors)
}
\arguments{
\item{vectors}{Matrix of row vectors}
}
\description{
Add an extra 0 dimension to vector
}
