% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate_biclustering}
\alias{annotate_biclustering}
\alias{annotate_biclustering,caclust-method}
\alias{annotate_biclustering,SingleCellExperiment-method}
\alias{annotate_cabinet}
\title{Perform gene overrepresentation analysis and annotate biclusters.}
\usage{
annotate_biclustering(
  obj,
  universe,
  org,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  ...
)

\S4method{annotate_biclustering}{caclust}(
  obj,
  universe,
  org,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  ...
)

\S4method{annotate_biclustering}{SingleCellExperiment}(
  obj,
  universe,
  org,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  ...,
  caclust_meta_name = "caclust"
)

\S4method{annotate_biclustering}{caclust}(
  obj,
  universe,
  org,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500,
  ...
)
}
\arguments{
\item{obj}{Either a `caclust` or `SingleCellExperiment` object.}

\item{universe}{All genes in data set.}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{alpha}{Adjusted p-value cutoff.
Only results with padj < alpha will be used.}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{...}{Further arguments.}

\item{caclust_meta_name}{Name under which the caclust object is stored in
the metadata of the SingleCellExperiment object.}
}
\value{
An object of type `caclust` with annotated biclusters.
}
\description{
Wrapper function for `per_cluster_goa` and `annotate_by_goa`.
}
\details{
`annotate_biclustering` performs per cluster GOA with a hypergeometric
 and annotates the biclustering results from CAbiNet.
}
