% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate_by_goa}
\alias{annotate_by_goa}
\title{Annotate CAbiNet results by gene overrepresentation
 analysis results.}
\usage{
annotate_by_goa(cabic, goa_res, alpha = 0.05)
}
\arguments{
\item{cabic}{Biclustering results of type `caclust`}

\item{goa_res}{List of goa results for each bicluster.}

\item{alpha}{Adjusted p-value cutoff.
Only results with padj < alpha will be used.}
}
\value{
Object of type `caclust` with annotated biclusters.
}
\description{
`annotate_by_goa` takes a biclustering results such as outputted by `caclust`
and annotates it with the gene overrepresentation analysis results (goa).

Conflicts between clusters that have the
 same highest ranking gene set are solved
 with the Hungarian/Munkres algorithm.
}
