% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{annotate_cabinet}
\alias{annotate_cabinet}
\title{Perform gene overrepresentation analysis and annotate biclusters.}
\usage{
annotate_cabinet(
  cabic,
  universe,
  org,
  set = "CellMarker",
  alpha = 0.05,
  min_size = 10,
  max_size = 500
)
}
\arguments{
\item{cabic}{A biclustering object of class "caclust"
as obtained from `caclust`.}

\item{universe}{All genes in data set.}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}

\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{alpha}{Adjusted p-value cutoff.
Only results with padj < alpha will be used.}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}
}
\value{
An object of type `caclust` with annotated biclusters.
}
\description{
Wrapper function for `per_cluster_goa` and `annotate_by_goa`.
}
\details{
`annotate_cabinet` performs per cluster GOA with a hypergeometric
 and annotates the biclustering results from CAbiNet.
}
