% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aR_metric}
\alias{aR_metric}
\title{calculates association ratio between to columns in a matrix.}
\usage{
aR_metric(matrix, origin, target)
}
\arguments{
\item{matrix}{a matrix}

\item{origin}{origin column (name or index)}

\item{target}{target column (name or index)}
}
\value{
Returns the association ratio between the columns from origin to target.
}
\description{
calculates association ratio between to columns in a matrix.
}
