% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.caclust}
\alias{as.caclust}
\alias{as.caclust,list-method}
\title{Try converting to caclust object from list or biclustlib.}
\usage{
as.caclust(obj)

\S4method{as.caclust}{list}(obj)
}
\arguments{
\item{obj}{A list with elements named same as caclust slots.}
}
\value{
A caclust object.
}
\description{
Converts object to caclust object if possible.
}
\details{
TODO
}
