% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{assign_cts}
\alias{assign_cts}
\title{Assign cell types to clusters using the Hungarian algorithm.}
\usage{
assign_cts(goa_res)
}
\arguments{
\item{goa_res}{List of goa results for each bicluster.}
}
\value{
A data frame with the assigned cell types and adjusted p-values.
}
\description{
Uses the hungarian algorithm (assignment problem)
to assign a cell type from the gene set overrepresentation
analysis to one (and only one) cluster.
}
