% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{biMAP}
\alias{biMAP}
\alias{biMAP,caclust-method}
\alias{biMAP,SingleCellExperiment-method}
\title{Compute biMAP}
\usage{
biMAP(obj, k = 30, rand_seed = 2358, method = "SNNdist", ...)

\S4method{biMAP}{caclust}(obj, k = 30, rand_seed = 2358, method = "SNNdist", ...)

\S4method{biMAP}{SingleCellExperiment}(
  obj,
  k = 30,
  rand_seed = 2358,
  method = "SNNdist",
  caclust_meta_name = "caclust",
  ...
)
}
\arguments{
\item{obj}{A caclust object or SingleCellExperiment object}

\item{k}{integer. Number of nearest neighbours to use to compute UMAP.}

\item{rand_seed}{integer. Random seed for UMAP.}

\item{method}{Can be either "SNNdist" or "spectral".}

\item{...}{Further arguments}

\item{caclust_meta_name}{The name of caclust object stored in metadata
(SingleCellExperiment object).}
}
\value{
A caclust object or SingleCellExperiment object.
}
\description{
The function takes either a \code{caclust} or \code{SingleCellExperiment}as input and
stores the biMAP in the "bimap" slot in the caclust object. If a
SingleCellExperiment was provided the caclust object is stored in its
metadata.
}
\details{
The biMAP cell and gene embeddings can be calculated via different methods
as controlled by the parameter \code{method}:
\itemize{
\item 'SNNdist'(Default): run UMAP on the distance matrix of cell-gene SNN graph
built up by caclust, which is '1-adj(SNN)'.
\item 'spectral': run UMAP on the selected eigenvectors of cell-gene graph
laplacian (only eligiable when algorithm is set as 'spectral' in 'caclust'
function)
}
}
