% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{biMAP_plotter}
\alias{biMAP_plotter}
\title{biMAP plotter}
\usage{
biMAP_plotter(
  caclust,
  color_by = "cluster",
  meta_df = NULL,
  type = "scatter",
  cell_size = 1,
  gene_size = 3,
  cell_alpha = 0.5,
  gene_alpha = 1,
  hex_n = 40,
  min_bin = 2,
  show_density = FALSE,
  contour_n = 5,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  colors = NULL,
  plotly = FALSE,
  max.overlaps = 12
)
}
\arguments{
\item{caclust}{caclust object with biMAP stored.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type 
(cell or gene) while "cluster" colors by the assigned cluster. If meta_df is
provided a column from the data.frame can be specified.}

\item{meta_df}{optional. data.frame that should have either gene or cell names
(or both) as rownames or a column named `name` and a column with the same name as 
`color_by`.}

\item{type}{Determines the kind of plot outputted. Either "scatter" 
(scatter plot), "contour" (density contour plot) or "hex" (binned hexagonals).}

\item{cell_size}{integer. Size of the cells in the plot.}

\item{gene_size}{integer. Size of the genes in the plot.}

\item{cell_alpha}{numeric. Alpha value of cells (between 0 and 1).}

\item{gene_alpha}{numeric. Alpha value of genes (between 0 and 1).}

\item{hex_n}{integer. Number of hexagonal bins that partition the x-axis.}

\item{min_bin}{integer. Only bins with more than `min_bin` cells are plotted.}

\item{show_density}{logical. Should the alpha value for plot type "hex" be
adjusted by the number of cells in the bin?}

\item{contour_n}{integer. Number of contour lines to plot.}

\item{color_genes}{logical. If TRUE colors genes by color_by.}

\item{label_groups}{logical. If TRUE puts the group label on the median 
coordinates of the point.}

\item{group_label_size}{integer. Size of the group label.}

\item{label_marker_genes}{logical or character vector. If TRUE, names of detected marker 
genes are displayed. If is character vector, the given marker genes are labeled.}

\item{colors}{Vector of hexadecimal color codes of the same length as the
categories defined by color_by or longer. If NULL ignored.}

\item{plotly}{If TRUE a hex plot with labels per hexbin will be made.}

\item{max.overlaps}{Exclude text labels that overlap too many things.}
}
\value{
biMAP plot as ggplot object.
}
\description{
This function plots the different variants of biMAP
}
