% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ca_biplot}
\alias{ca_biplot}
\alias{ca_biplot,cacomp-method}
\title{#' Plot of 2D CA projection of the data.}
\usage{
ca_biplot(
  obj,
  metadf = NULL,
  color_by = NULL,
  xdim = 1,
  ydim = 2,
  princ_coords = 1,
  row_labels = NULL,
  col_labels = NULL,
  type = "plotly",
  ...
)

\S4method{ca_biplot}{cacomp}(
  obj,
  metadf = NULL,
  color_by = NULL,
  xdim = 1,
  ydim = 2,
  coords = 1,
  row_labels = NULL,
  col_labels = NULL,
  type = "plotly",
  rm.show = TRUE,
  ...,
  caclust = NULL
)
}
\description{
#' Plot of 2D CA projection of the data.

it to accept caclust-class as imput and color the genes and cells in a 2D plot
}
