% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAclust.R
\docType{package}
\name{CAbiNet}
\alias{CAbiNet}
\title{CAbiNet: Correspondence Analysis for Biclustering on Networks}
\description{
CAbiNet provides functions to build graphs of cells and genes as well as
graph clustering methods appropriate for such graphs. Cells and genes can be
co-embedded in a biMAP.
}
\section{Mypackage functions}{

The mypackage functions ...
}

