% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felines.R
\name{cabinet_cat}
\alias{cabinet_cat}
\title{Check For A Cabinet Office Cat}
\usage{
cabinet_cat(name)
}
\arguments{
\item{name}{A cat name provided as a character string.}
}
\description{
Check whether a name belongs to one of the cats from the UK
    government's Cabinet Office.
}
\examples{
cabinet_cat("Evie")
}
