% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAclust.R
\docType{package}
\name{CAclust}
\alias{CAclust}
\title{CAclust: Biclustering in CA space for single-cell RNA-seq data}
\description{
CAclust provides functions to build graphs of cells and genes as well as
graph clustering methods appropriate for such graphs.
}
\section{Mypackage functions}{

The mypackage functions ...
}

