% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{caclust}
\alias{caclust}
\alias{caclust,caclust-method}
\alias{caclust,SingleCellExperiment-method}
\title{caclust}
\usage{
caclust(
  obj,
  k_sym,
  k_asym = k_sym,
  cacomp_meta_name = "caobj",
  caclust_meta_name = "caclust",
  ...
)

\S4method{caclust}{caclust}(
  obj,
  k_sym,
  k_asym = k_sym,
  cacomp_meta_name = "caobj",
  caclust_meta_name = "caclust",
  ...
)

\S4method{caclust}{SingleCellExperiment}(
  obj,
  k_sym,
  k_asym = k_sym,
  cacomp_meta_name = "caobj",
  caclust_meta_name = "caclust",
  ...
)
}
\arguments{
\item{obj}{SingleCellExperiment object}

\item{k_sym}{neighour of nearest neighours, see run_caclust  function}

\item{cacomp_meta_name}{the name of cacomp object stored in metadata(SingleCellExperiment object)}

\item{caclust_meta_name}{the name of caclust object stored in metadata(SingleCellExperiment object)}
}
\value{
an caclust object or SingleCellExperiment objects
}
\description{
run biclustering algorithm, compatible for matrix and sce input
}
\details{
TODO
}
