% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_distances}
\alias{calc_distances}
\title{Calculate distances for bigraph}
\usage{
calc_distances(caobj)
}
\arguments{
\item{caobj}{A cacomp object with principal and standard coordinates calculated.}
}
\value{
List with elements:
\itemize{
\item "cc": cell-cell euclidean distances
\item "gg": gene-gene euclidean distances
\item "cg": cell-gene association ratio
\item "gc": gene-cell association ratio
}
}
\description{
Calculate distances for bigraph
}
