% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_euclidean}
\alias{calc_euclidean}
\title{Calculate euclidean distance}
\usage{
calc_euclidean(mat)
}
\arguments{
\item{mat}{matrix.}
}
\value{
Matrix with pairwise euclidean distances.
}
\description{
Calculates euclidean distance.
}
