% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_overlap}
\alias{calc_overlap}
\title{WIP replacement for `determine_overlap` function}
\usage{
calc_overlap(cc_adj, cg_adj)
}
\arguments{
\item{cc_adj}{cell-cell adjacency matrix}

\item{cg_adj}{cell-gene adjacency matrix}
}
\description{
DO NOT USE! Slower than R implementation
}
