% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellmarker_v2}
\alias{cellmarker_v2}
\title{CellMarker V2}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 96074 rows and 20 columns.
}
\usage{
cellmarker_v2
}
\description{
All CellMarker V2 obtained from: http://bio-bigdata.hrbmu.edu.cn/CellMarker/CellMarker_download_files/file/Cell_marker_All.xlsx
Unprocessed data.
}
\keyword{datasets}
