% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_caclust}
\alias{check_caclust}
\title{Checks if caclust-class was constructed correctly}
\usage{
check_caclust(object)
}
\arguments{
\item{object}{caclust object}
}
\value{
If object is a valid caclust object returns TRUE, otherwise the errors.
}
\description{
Checks if caclust-class was constructed correctly
}
