% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{contour_plot}
\alias{contour_plot}
\title{TODO}
\usage{
contour_plot(
  umap_cells,
  umap_genes,
  color_by,
  color_genes,
  colors,
  interact_genes,
  cell_size,
  gene_size,
  gene_alpha,
  contour_n
)
}
\description{
TODO
}
