% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{convert_to_biclust}
\alias{convert_to_biclust}
\title{Converts CAclust results to biclustlib results}
\usage{
convert_to_biclust(caclust)
}
\arguments{
\item{caclust}{A caclust object}
}
\value{
An object of type "Biclust".
}
\description{
Converts CAclust results to biclustlib results
}
