% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_bigraph}
\alias{create_bigraph}
\title{Combine kNN graphs to large cell-gene adjecency matrix}
\usage{
create_bigraph(
  caobj,
  k_c,
  k_g,
  k_cg,
  k_gc,
  loops = FALSE,
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE,
  marker_genes = NULL,
  method = BiocNeighbors::KmknnParam(),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{caobj}{A cacomp object with standard and principal coordinates
calculated.}

\item{k_c}{k for cell-cell kNN, integer.}

\item{k_g}{k for gene-gene kNN, integer.}

\item{k_cg}{k for cell-gene kNN, integer.}

\item{k_gc}{k for gene-cell kNN, interger.}

\item{loops}{TRUE/FALSE. If TRUE self-loops are allowed, otherwise not.}

\item{select_genes}{TRUE/FALSE. Should genes be selected by whether they have
an edge in the cell-gene kNN graph?}

\item{prune_overlap}{TRUE/FALSE. If TRUE edges to genes that share less
than \code{overlap} of genes with the nearest neighbours of the cell are removed.
Pruning is only performed if select_genes = TRUE.}

\item{overlap}{Numeric between 0 and 1. Overlap cutoff applied if
prune_overlap = TRUE.}

\item{calc_gene_cell_kNN}{TRUE/FALSE. If TRUE a cell-gene graph is calculated
by choosing the \code{k_gc} nearest cells for each gene. If FALSE the cell-gene
graph is transposed.}

\item{marker_genes}{character. Optional. Names of known marker genes that
should be excempt from any pruning on the graph and be kept.}

\item{method}{\link[BiocNeighbors:BiocNeighborParam]{BiocNeighbors::BiocNeighborParam} object specifying the
algorithm to use. see Details.}

\item{BPPARAM}{\link{BiocParallel} settings parameter. By default single core
\code{\link[BiocParallel:SerialParam-class]{BiocParallel::SerialParam()}} but other parameters can be passed.}
}
\value{
Adjacency matrix of type \code{dgCMatrix}.
The combined adjacency matrix consists of the cell-cell graph, gene-gene
graph and cell-gene/gene-cell graph.
}
\description{
Builds a single adjacency matrix consisting of cells and genes from 4
seperate sub kNN-graphs.
}
\details{
\code{method} should be a kNN algorithm defined by
\link[BiocNeighbors:BiocNeighborParam]{BiocNeighbors::BiocNeighborParam}. For exact kNN search use
\code{BiocNeighbors::KmknnParam()} or \code{BiocNeighbors::VptreeParam()}.
}
