% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{create_SNN}
\alias{create_SNN}
\title{Create SNN graph from caobj}
\usage{
create_SNN(
  caobj,
  k_c,
  k_g,
  k_cg,
  k_gc,
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE
)
}
\arguments{
\item{caobj}{A cacomp object with standard and principal coordinates 
calculated.}

\item{k_c}{k for cell-cell kNN}

\item{k_g}{k for gene-gene kNN}

\item{k_cg}{k for cell-gene kNN}

\item{k_gc}{k for gene-cell kNN}

\item{select_genes}{TRUE/FALSE. Should genes be selected by wether they have
an edge in the cell-gene kNN graph?}

\item{prune_overlap}{TRUE/FALSE. If TRUE edges to genes that share less
than `overlap` of genes with the nearest neighbours of the cell are removed.}

\item{overlap}{Numeric between 0 and 1. Overlap cutoff if
prune_overlap = TRUE.}

\item{calc_cell_gene_kNN}{TRUE/FALSE. If TRUE a cell-gene graph is calculated
by choosing the `k_gc` nearest cells for each gene. If FALSE the cell-gene
graph is transposed.}
}
\value{
A sparse adjacency Matrix of type "dgCMatrix". The values in the matrix
are the Jaccard similarity between nodes in the graph. The range between 0
and 1, with 0 meaning that no edges are shared between nodes, wheras 1 means 
all edges are shared between nodes.
}
\description{

}
