% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{eigengap}
\alias{eigengap}
\title{Detecting eigengap.}
\usage{
eigengap(e, v)
}
\arguments{
\item{e}{Eigenvalues of type 'vector'}

\item{v}{Eigenvector matirx of type 'matrix/array' whose columns are single eigenvectors.}
}
\value{
The selected most important eigenvectors for downstream clustering, of type 'matrix/array'.
}
\description{
Searching for the largest gap between conjuncted eigenvalues which are sorted 
in an ascending order, the most important eigenvalues and eigenvectors are the ones 
before the detected eigengap.
}
