% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{feature_biMAP}
\alias{feature_biMAP}
\alias{feature_biMAP,SingleCellExperiment,caclust-method}
\alias{feature_biMAP,SingleCellExperiment,ANY-method}
\title{BiMAP visualzation of expression of features}
\usage{
feature_biMAP(
  obj,
  caclust = NULL,
  biMAP_meta_name = "biMAP_SNNdist",
  feature = NULL,
  color_cells_by = "expression",
  assay = "logcounts",
  ...
)

\S4method{feature_biMAP}{SingleCellExperiment,caclust}(
  obj,
  caclust = NULL,
  biMAP_meta_name = "biMAP_SNNdist",
  feature = NULL,
  color_cells_by = "expression",
  assay = "logcounts",
  ...
)

\S4method{feature_biMAP}{SingleCellExperiment,ANY}(
  obj,
  caclust,
  caclust_meta_name = "caclust",
  feature = NULL,
  color_cells_by = "expression",
  assay = "logcounts",
  ...
)
}
\arguments{
\item{obj}{SinleCellExperiment object}

\item{caclust}{caclust object with bimap slot}

\item{biMAP_meta_name}{character. Name of biMAP coordiantes data.frame stored in metadata(sce obj)}

\item{feature}{character. Name of gene to visualize}

\item{color_cells_by}{character. Default: expression}

\item{assay}{character. Name of assay in SingleCellExperiment used for visualization.}
}
\description{
BiMAP visualzation of expression of features
}
