% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{feature_biUMAP}
\alias{feature_biUMAP}
\title{plot biUMAP with gene expression}
\usage{
feature_biUMAP(
  umap_coords,
  sce,
  feature = NULL,
  color_cells_by = "expression",
  assay = "logcounts"
)
}
\arguments{
\item{assay}{}
}
\description{
plot biUMAP with gene expression
}
