% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{filter_gene_sets}
\alias{filter_gene_sets}
\title{Filter gene sets by size}
\usage{
filter_gene_sets(gene_sets, min_size = 10, max_size = 500)
}
\arguments{
\item{gene_sets}{A named list of gene sets (name is the gene set,
each element of the list contains a vector of genes)}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}
}
\value{
Filtered list of gene sets.
}
\description{
This function cuts removes gene sets below and above a certain size.
}
