% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{format_gene_sets}
\alias{format_gene_sets}
\title{Changes long format gene set data frame into a named list of gene sets.}
\usage{
format_gene_sets(gene_sets)
}
\arguments{
\item{gene_sets}{data frame with first column gene set names and the
second column the gene symbol. Long format expected.}
}
\value{
A named list with gene sets as names and genes as vectors.
}
\description{
Takes in a data frame with 2 columns (gene set name and gene name)
and accumulates all genes in the same gene set into a list element.
}
