% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_eigen}
\alias{get_eigen}
\title{Get eigenvectors from spectral clustering}
\usage{
get_eigen(object)
}
\arguments{
\item{object}{caclust object}
}
\description{
Get eigenvectors from spectral clustering
}
