% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_majority}
\alias{get_majority}
\title{Determines majority in a vector}
\usage{
get_majority(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single element of x that has the highest count.
}
\description{
Changed version of mclust::majorityVote. Ties are broken randomly.
}
