% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_ncells}
\alias{get_ncells}
\title{Get the number of cells from each cluster per hexbin}
\usage{
get_ncells(umap_cells)
}
\arguments{
\item{umap_cells}{data frame containing the umap coordinates of cells with
the hex saved in hexbin}
}
\value{
bimap_coordinates of cells for cells with the makeup of each hexbin in the column
prop.
}
\description{
Get the number of cells from each cluster per hexbin
}
