% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_params}
\alias{get_params}
\title{Get parameters for generating clustering}
\usage{
get_params(object)
}
\arguments{
\item{object}{caclust object}
}
\description{
Get parameters for generating clustering
}
