% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_per_gene_overlap}
\alias{get_per_gene_overlap}
\title{Calculates the overlap of chosen genes between neighbouring cells}
\usage{
get_per_gene_overlap(idx, cg_adj, cc_adj)
}
\arguments{
\item{idx}{indices of cell of interest}

\item{cg_adj}{cell-gene adjacency matrix}

\item{cc_adj}{cell-cell adjacency matrix}
}
\value{
numeric vector indicating the overlap between neighbours for each gene that
is a neighbour in `cg_adj`. Same length as number of nearest neighbour genes
in `cg_adj` for each cell.
}
\description{
For the cell indicated with `idx` in the rows of `cc_adj` and `cg_adj`,
the genes and cells with and edge to the chosen cell are determined.
Among the connected it is then determined how many have an edge to the same
genes as the chosen cell. The overlap (cells with edge to same gene)/
(Nr. of cells) is returned.
}
