% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hex_plot}
\alias{hex_plot}
\title{TODO}
\usage{
hex_plot(
  umap_cells,
  umap_genes,
  color_by,
  colors,
  color_genes,
  hex_n,
  min_bin,
  show_density,
  cell_alpha,
  gene_size,
  gene_alpha,
  interact_genes
)
}
\description{
TODO
}
