% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{indx_to_spmat}
\alias{indx_to_spmat}
\title{Convert index matrix to sparse matrix.}
\usage{
indx_to_spmat(indx_mat, row_names, col_names)
}
\arguments{
\item{indx_mat}{row-wise index matrix. Usually a kNN graph.}

\item{row_names}{the rownames of the resulting sparse matrix.}

\item{col_names}{the column names that correspond to the indexes in indx_mat.}
}
\value{
a `Matrix::sparseMatrix` object.
}
\description{
Convert index matrix to sparse matrix.
}
