% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/helpers.R
\name{is.empty}
\alias{is.empty}
\title{Helper function to check if object is empty.}
\usage{
is.empty(x)

is.empty(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE if x has length 0 and is not NULL. FALSE otherwise

TRUE if x has length 0 and is not NULL. FALSE otherwise
}
\description{
Helper function to check if object is empty.

Helper function to check if object is empty.
}
