% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{load_gene_set}
\alias{load_gene_set}
\title{Load the required gene set.}
\usage{
load_gene_set(set = "CellMarker", org)
}
\arguments{
\item{set}{Name of the gene set. Currently only supports "CellMarker"}

\item{org}{Short name of the organism. "mm" for mouse, "hs" for human.}
}
\value{
data frame with columns "cell_type" and "gene".
}
\description{
Loads the speciefied gene set and subsets to the required organism.
}
