% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{metadata_biMAP}
\alias{metadata_biMAP}
\title{Plot continous or categorical data as a biMAP}
\usage{
metadata_biMAP(caclust, meta_df, color_cells_by)
}
\arguments{
\item{caclust}{A caclust object with bimap coordinates stored.}

\item{meta_df}{data.frame with column specified by color_cells_by with cell
names as rownames (e.g. colData() from SingleCellExperiment object)}

\item{color_cells_by}{coloumn name in meta_df}
}
\value{
ggplot of metadata biMAP
}
\description{
Plot continous or categorical data as a biMAP
}
