% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mix_rgb}
\alias{mix_rgb}
\title{Mixes the colors of two clusters proportionally.}
\usage{
mix_rgb(df, colors, cell, color_by)
}
\arguments{
\item{df}{data.frame of cells with clusters in `color_by` and assigned 
hex bin in `hexbin`.}

\item{colors}{colors to be mixed.}

\item{cell}{Which hexbin to mix colors in.}

\item{color_by}{Column name where the clusters/groups are stored in `df`.}
}
\value{
Mixed color as hex code.
}
\description{
Mixes the colors of two clusters proportionally.
}
