% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_caclust}
\alias{new_caclust}
\title{Create new caclust object}
\usage{
new_caclust(...)
}
\arguments{
\item{...}{arguments forwarded to new. Should be named slots of a caclust
object}
}
\description{
TODO
}
