% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{NormLaplacian}
\alias{NormLaplacian}
\title{Calculate Normalized Graph Laplacian}
\usage{
NormLaplacian(adj)
}
\arguments{
\item{adj}{The adjacency matix of type 'matrix/array'.}
}
\value{
A sparsed normalized graph laplacian matrix of type "dgCmatrix".
}
\description{
Calculate Normalized graph laplacian of the input adjacency matrix, the graph laplacian
L = I-D^(-1/2)AD^(-1/2), where D is a diagonal matrix with row sums of A as entries.
}
