% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{optimal_km}
\alias{optimal_km}
\title{An integration of several runs of kmens with different random seeds}
\usage{
optimal_km(x, k, num_seeds = 10, iter_max = 10, ...)
}
\arguments{
\item{x}{Matrix. This function will cluster the rows of the input matrix.}

\item{k}{Integer. Number of cluters to detect for kmeans.}

\item{num_seeds}{Integer. Number of trials with random seeds}

\item{iter_max}{Integer. Number of iterations.}

\item{...}{Further arguments handed to stats::kmeans}
}
\value{
Returns an object of class "kmeans" with is a list with several components
(see ?kmeans) which guves the local optimal kmeans clustering result within
#num_seeds trials.
}
\description{
This function will select the optimal clustering result from several kmeans runs
with different random seeds, The clustering result with the smallest
within-cluster-sum of squared distances will be selected.
}
