% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{optimal_skm}
\alias{optimal_skm}
\title{An integration of several runs of skmens with different random seeds}
\usage{
optimal_skm(
  x,
  k,
  num_seeds = 10,
  method = NULL,
  m = 1,
  weights = 1,
  control = list(),
  ...
)
}
\arguments{
\item{x}{Matrix. This function will cluster the rows of the input matrix.}

\item{k}{Integer. Number of cluters to detect for skmeans.}

\item{num_seeds}{Integer. Number of trials with random seeds}

\item{method}{a character string specifying one of the built-in
    methods for computing spherical \eqn{k}-means partitions, or a
    function to be taken as a user-defined method, or \code{NULL}
    (default value).  If a character string, its lower-cased version is
    matched against the lower-cased names of the available built-in
    methods using \code{\link{pmatch}}.  See \bold{Details} for
    available built-in methods and defaults.}

\item{m}{a number not less than 1 controlling the softness of the
    partition (as the \dQuote{fuzzification parameter} of the fuzzy
    \eqn{c}-means algorithm).  The default value of 1 corresponds to
    hard partitions; values greater than one give partitions of
    increasing softness obtained from a generalized soft spherical
    \eqn{k}-means problem.}

\item{weights}{a numeric vector of non-negative case weights.
    Recycled to the number of objects given by \code{x} if necessary.}

\item{control}{a list of control parameters.  See \bold{Details}.}
}
\value{
Returns an object inheriting from classes skmeans and pclust (see ?skmeans) which 
gives the local optimal skmeans clustering result within several trials.
}
\description{
This function will select the optimal clustering result from several 
skmeans::skmeans runs with different random seeds, The clustering result with the smallest 
within-cluster-sum of squared distances will be selected.
}
