% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{parse_ratio}
\alias{parse_ratio}
\title{turns string of a ratio in a number.}
\usage{
parse_ratio(ratio)
}
\arguments{
\item{ratio}{a string of the form "1/2"}
}
\value{
The numeric value the string represented.
}
\description{
Adapted from DOSE::parse_ratio.
}
