% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{perform_goa}
\alias{perform_goa}
\title{Adapted from `DOSE:::enricher_interal`.
Performs Gene Overrepresentation Analysis.}
\usage{
perform_goa(gois, gene_sets, universe, min_size, max_size, verbose = TRUE)
}
\arguments{
\item{gois}{Genes of interest. Usually the co-clustered genes.}

\item{gene_sets}{Named list of gene sets and their genes.}

\item{universe}{All genes in data set.}

\item{min_size}{Min. number of genes in the gene set.}

\item{max_size}{Max number of genes in the gene set.
Set to Inf if you want to keep all genes.}

\item{verbose}{Toggles verbosity of warnings.}
}
\value{
A data frame with the results of the go-analysis.
}
\description{
perform_goa takes a number of genes of interest (gois)
and a list of gene sets and performs gene overrepresentation
analysis.
}
\details{
perform_goa performs a hypergeometric test on the gois and gene sets.
Gene sets are trimmed to genes that are present in universe
 (usually all the genes available in the data set) and trimmed to set
 min and max size.
}
