% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contour_biMAP}
\alias{plot_contour_biMAP}
\alias{plot_contour_biMAP,caclust-method}
\title{Plot contour biMAP}
\usage{
plot_contour_biMAP(
  obj,
  meta_df = NULL,
  color_by = "cluster",
  gene_size = 2,
  gene_alpha = 0.8,
  contour_n = 5,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  ...
)

\S4method{plot_contour_biMAP}{caclust}(
  obj,
  meta_df = NULL,
  color_by = "cluster",
  gene_size = 2,
  gene_alpha = 0.8,
  contour_n = 5,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  ...
)
}
\arguments{
\item{obj}{Object that stores biMAP coordinates. Can be either a "caclust"
object or of class "SingleCellExperiment".}

\item{meta_df}{optional. data.frame that should have either gene or cell names
(or both) as rownames or a column named `name` and a column with the same name as 
`color_by`.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type 
(cell or gene) while "cluster" colors by the assigned cluster. If meta_df is
provided a column from the data.frame can be specified.}

\item{gene_size}{integer. Size of the genes in the plot.}

\item{gene_alpha}{numeric. Alpha value of genes (between 0 and 1).}

\item{contour_n}{integer. Number of contour lines to plot.}

\item{color_genes}{logical. If TRUE colors genes by color_by.}

\item{label_groups}{logical. If TRUE puts the group label on the median 
coordinates of the point.}

\item{group_label_size}{integer. Size of the group label.}

\item{label_marker_genes}{logical. If TRUE, names of genes are displayed.}

\item{...}{Further arguments.}
}
\value{
contour biMAP plot as ggplot object.
}
\description{
Cell density is plotted as contour lines and genes are plotted as points
over the contours.
}
\seealso{
Other plot_*_biMAP: 
\code{\link{plot_hex_biMAP}()},
\code{\link{plot_scatter_biMAP}()}
}
\concept{plot_*_biMAP}
