% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_feature_biMAP}
\alias{plot_feature_biMAP}
\title{plot biMAP with gene expression}
\usage{
plot_feature_biMAP(
  sce,
  caclust,
  feature = NULL,
  color_cells_by = "expression",
  assay = "logcounts"
)
}
\arguments{
\item{sce}{SinleCellExperiment object}

\item{caclust}{caclust object}

\item{feature}{character. Name of gene to visualize}

\item{color_cells_by}{character. Default: expression}

\item{assay}{character. Name of assay in SingleCellExperiment used for visualization.}
}
\description{
plot biMAP with gene expression
}
