% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_anno.R
\name{plot_goa_res}
\alias{plot_goa_res}
\title{Plot gene overrepresentation analysis results.}
\usage{
plot_goa_res(goa_res, nres = 3)
}
\arguments{
\item{goa_res}{List of goa results for each cluster.}

\item{nres}{Number of top gene sets to plot.}
}
\value{
A ggplot2 object. Dotplot of the top `nres` gene sets per cluster.
}
\description{
Plots the gene overrepresentation analysis results for each cluster.
}
