% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metadata_biMAP}
\alias{plot_metadata_biMAP}
\alias{plot_metadata_biMAP,caclust-method}
\alias{plot_metadata_biMAP,SingleCellExperiment-method}
\title{Plot continous or categorical data as a biMAP}
\usage{
plot_metadata_biMAP(obj, color_cells_by, ...)

\S4method{plot_metadata_biMAP}{caclust}(obj, color_cells_by, ..., meta_df)

\S4method{plot_metadata_biMAP}{SingleCellExperiment}(obj, color_cells_by, ..., caclust_meta_name = "caclust")
}
\arguments{
\item{obj}{caclust or SingleCellExperiment object.}

\item{color_cells_by}{coloumn name in meta_df}

\item{...}{Further arguments.}

\item{meta_df}{data.frame with column specified by color_cells_by with cell
names as rownames (e.g. colData() from SingleCellExperiment object)}

\item{caclust_meta_name}{Name under which the caclust object is stored in 
the metadata of the SingleCellExperiment object.}
}
\value{
ggplot of metadata biMAP
}
\description{
Plot continous or categorical data as a biMAP
}
