% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scatter_biMAP}
\alias{plot_scatter_biMAP}
\alias{plot_scatter_biMAP,caclust-method}
\alias{plot_scatter_biMAP,SingleCellExperiment-method}
\alias{plot_biMAP}
\title{Plots a biMAP scatter plot}
\usage{
plot_scatter_biMAP(
  obj,
  meta_df = NULL,
  color_by = "cluster",
  cell_size = 1,
  gene_size = 2,
  cell_alpha = 0.7,
  gene_alpha = 0.8,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  interactive = FALSE,
  ...
)

\S4method{plot_scatter_biMAP}{caclust}(
  obj,
  meta_df = NULL,
  color_by = "cluster",
  cell_size = 1,
  gene_size = 2,
  cell_alpha = 0.7,
  gene_alpha = 0.8,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  interactive = FALSE,
  ...
)

\S4method{plot_scatter_biMAP}{SingleCellExperiment}(
  obj,
  meta_df = NULL,
  color_by = "cluster",
  cell_size = 1,
  gene_size = 2,
  cell_alpha = 0.7,
  gene_alpha = 0.8,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  interactive = FALSE,
  ...,
  caclust_meta_name = "caclust"
)

plot_biMAP(
  obj,
  meta_df = NULL,
  color_by = "cluster",
  cell_size = 1,
  gene_size = 2,
  cell_alpha = 0.7,
  gene_alpha = 0.8,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  label_marker_genes = FALSE,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{obj}{Object that stores biMAP coordinates. Can be either a "caclust"
object or of class "SingleCellExperiment".}

\item{meta_df}{optional. data.frame that should have either gene or cell names
(or both) as rownames or a column named `name` and a column with the same name as 
`color_by`.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type 
(cell or gene) while "cluster" colors by the assigned cluster. If meta_df is
provided a column from the data.frame can be specified.}

\item{cell_size}{integer. Size of the cells in the plot.}

\item{gene_size}{integer. Size of the genes in the plot.}

\item{cell_alpha}{numeric. Alpha value of cells (between 0 and 1).}

\item{gene_alpha}{numeric. Alpha value of genes (between 0 and 1).}

\item{color_genes}{logical. If TRUE colors genes by color_by.}

\item{label_groups}{logical. If TRUE puts the group label on the median 
coordinates of the point.}

\item{group_label_size}{integer. Size of the group label.}

\item{label_marker_genes}{logical. If TRUE, names of genes are displayed.}

\item{interactive}{If TRUE plot_scatter_biMAP returns an interactive (html) 
plot. If FALSE a ggplot object is returned.}

\item{...}{Further arguments.}

\item{caclust_meta_name}{Name under which the caclust object is stored in 
the metadata of the SingleCellExperiment object.}
}
\value{
biMAP scatter plot as ggplot object.
}
\description{
Plots cells and genes as points. Genes are plotted on top of the cells in
a larger size to make visual differentiation easier.
}
\seealso{
Other plot_*_biMAP: 
\code{\link{plot_contour_biMAP}()},
\code{\link{plot_hex_biMAP}()}
}
\concept{plot_*_biMAP}
