% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rm_monoclusters}
\alias{rm_monoclusters}
\title{Remove clusters only consisting of cells/genes}
\usage{
rm_monoclusters(bic)
}
\arguments{
\item{bic}{object of class biclust}
}
\value{
biclust object with monoclusters removed.
}
\description{
Takes an object of class biclust and removes all clusters that only consist
of cells or genes.
}
