% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{run_biMAP}
\alias{run_biMAP}
\title{Run UMAP embedding for cell-gene graph built up by caclust.}
\usage{
run_biMAP(
  obj,
  caobj = NULL,
  k = 30,
  rand_seed = 2358,
  method = "SNNdist",
  use_SNN = TRUE,
  features = NULL
)
}
\arguments{
\item{obj}{results from biclustering of class "caclust"}

\item{caobj}{A cacomp object with principal and standard coordinates
calculated. Only needs to be supplied when using method "ca".}

\item{k}{integer. Number of nearest neighbours to use to compute UMAP.}

\item{rand_seed}{integer. Random seed for UMAP.}

\item{method}{Can be either "SNNdist", "spectral" or "ca". When using "ca",
a "cacomp" object has to be provided for \code{caobj}.}

\item{use_SNN}{TRUE/FALSE. This parameter only works when method == 'ca'.
If TRUE, it will calculate bimap embedding of genes and cells contained in the SNN graph together with the given 'fearures' if any.
If FALSE, only calculate bimap embedding of cells contained in the SNN graph together with the given 'fearures' if any.}
}
\value{
caclust object with biMAP coordinates stored in the \code{bimap} slot.
}
\description{
This function takes cacomp and caclust object as input to calculate UMAP embedding
of cell-gene graph in several different ways:
\itemize{
\item 'SNNdist'(Default): run UMAP on the distance matrix of cell-gene SNN graph
built up by caclust, which is '1-adj(SNN)'.
\item 'spectral': run UMAP on the selected eigenvectors of cell-gene graph
laplacian (only eligiable when algorithm is set as 'spectral' in 'caclust'
function)
\item 'ca': run UMAP on the singular vectors from Correspondence Analysis.
}
}
