% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_biUMAP}
\alias{run_biUMAP}
\title{Plots UMAP depicting both cells and genes.}
\usage{
run_biUMAP(caobj, caclust_obj, k_umap, rand_seed = 2358, algorithm = "leiden")
}
\arguments{
\item{caobj}{A cacomp object with principal and standard coordinates 
calculated.}

\item{caclust_obj}{results from biclustering of class "caclust"}

\item{k_umap}{Number of nearest neighbours to use from the SNN graph for
UMAP}

\item{rand_seed}{Random seed for UMAP.}
}
\value{
data frame containing the UMAP coordinates of cells and genes.
}
\description{
TODO
}
