% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_biUMAP_leiden}
\alias{run_biUMAP_leiden}
\title{Plots UMAP depicting both cells and genes.}
\usage{
run_biUMAP_leiden(caobj, SNN, k_umap, rand_seed = 2358)
}
\arguments{
\item{caobj}{A cacomp object with principal and standard coordinates 
calculated.}

\item{SNN}{dense or sparse matrix. A SNN graph.}

\item{k_umap}{Number of nearest neighbours to use from the SNN graph for
UMAP}

\item{rand_seed}{Random seed for UMAP.}
}
\value{
data frame containing the UMAP coordinates of cells and genes.
}
\description{
TODO
}
