% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{run_plot_biMAP}
\alias{run_plot_biMAP}
\title{Plot biMAP}
\usage{
run_plot_biMAP(
  umap_coords,
  metadata = NULL,
  color_by = "cluster",
  type = "scatter",
  cell_size = 1,
  gene_size = 3,
  hex_n = 40,
  min_bin = 2,
  contour_n = 5,
  cell_alpha = 0.5,
  gene_alpha = 1,
  show_density = FALSE,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  labels_per_group = 1,
  show_gene_meta = TRUE,
  label_marker_gene = FALSE
)
}
\arguments{
\item{umap_coords}{data frame as outputted by `run_biMAP_*`}

\item{metadata}{optional. data.frame that should have either gene or cell names
(or both) in as rownames or a column named `name` and a column with the same name as 
`color_by`.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type 
(cell or gene) while "cluster" colors by the assigned cluster.}

\item{type}{Either "scatter", "contour" or "hex".}
}
\value{
ggplot of UMAP
}
\description{
Plot biMAP
}
