% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{TODO}
\usage{
scatter_plot(
  umap_cells,
  umap_genes,
  color_by,
  colors,
  color_genes,
  cell_size,
  cell_alpha,
  gene_size,
  gene_alpha,
  interact_genes,
  interact_cells
)
}
\description{
TODO
}
