% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show.caclust}
\alias{show.caclust}
\alias{show,caclust-method}
\title{Print caclust object in console}
\usage{
show.caclust(object)

\S4method{show}{caclust}(object)
}
\arguments{
\item{object}{a caclust object}
}
\description{
Print caclust object in console
}
