% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tbl_paste}
\alias{tbl_paste}
\title{concatenate table() results for hex_plot.}
\usage{
tbl_paste(x)
}
\arguments{
\item{x}{a vector}
}
\value{
concatenated results for the table.
}
\description{
concatenate table() results for hex_plot.
}
