% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bicplot}
\alias{bicplot}
\alias{bicplot,caclust-method}
\alias{bicplot,SingleCellExperiment-method}
\title{Plot of 2D CA projection of the clustering results.}
\usage{
bicplot(
  obj,
  xdim = 1,
  ydim = 2,
  princ_coords = 1,
  row_labels = NULL,
  col_labels = NULL,
  type = "ggplot",
  show_all = TRUE,
  show_rm = FALSE,
  ...
)

\S4method{bicplot}{caclust}(
  obj,
  xdim = 1,
  ydim = 2,
  princ_coords = 1,
  row_labels = NULL,
  col_labels = NULL,
  type = "ggplot",
  show_all = TRUE,
  show_rm = FALSE,
  ...,
  caobj
)

\S4method{bicplot}{SingleCellExperiment}(
  obj,
  xdim = 1,
  ydim = 2,
  princ_coords = 1,
  row_labels = NULL,
  col_labels = NULL,
  type = "ggplot",
  show_all = TRUE,
  show_rm = FALSE,
  ...,
  caclust_meta_name = "caclust",
  cacomp_meta_name = "CA"
)
}
\arguments{
\item{obj}{caclust or SingleCellExperiment object containing clustering 
results. The SingleCellExperiment should also have a CA dimensional reduction.}

\item{xdim}{Integer. The dimension for the x-axis. Default 1.}

\item{ydim}{Integer. The dimension for the y-axis. Default 2.}

\item{princ_coords}{Integer. If 1 then principal coordinates are used for 
the rows, if 2 for the columns. Default 1 (rows).}

\item{row_labels}{Numeric vector. Indices for the rows for which a label 
should be added
(label should be stored in rownames). Default NULL.}

\item{col_labels}{Numeric vector. Indices for the columns for which a label 
should be added
(label should be stored in colnames).
Default NULL (no columns).}

\item{type}{String. Type of plot to draw. Either "ggplot" or "plotly". 
Default "ggplot".}

\item{show_all}{logical. If FALSE cells/genes that are not in col_metadata/
row_metadata are not plotted. If *_metadata is NULL, the cell or genes 
respectively will still be plotted.}

\item{show_rm}{Show pruned genes during biclustering.}

\item{...}{Further arguments.}

\item{caobj}{An object of class "cacomp" with the relevant standardized and 
principal coordinates calculated.}

\item{caclust_meta_name}{the name of caclust object stored in 
metadata(SingleCellExperiment object). Default: 'caclust.'}

\item{cacomp_meta_name}{Character. The name of cacomp object stored in 
metadata(SingleCellExperiment object). Default: 'caobj'.}
}
\value{
Plot of class "plotly" or "ggplot".
}
\description{
Plots the first 2 dimensions of the rows and columns in the same plot.
}
\details{
Choosing type "plotly" will generate an interactive html plot with the 
package plotly.
Type "ggplot" generates a static plot.
Depending on whether `princ_coords` is set to 1 or 2 either
the principal coordinates of either the rows (1) or the columns (2)
are chosen. For the other the standard coordinates are plotted 
(assymetric biplot).
Labels for rows and columns should be stored in the row and column names 
respectively.
}
