% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{biMAP}
\alias{biMAP}
\alias{biMAP,cacomp,caclust-method}
\alias{biMAP,SingleCellExperiment,ANY-method}
\title{biMAP}
\usage{
biMAP(
  obj,
  caclust_obj,
  k_umap,
  cacomp_meta_name = "caobj",
  caclust_meta_name = "caclust",
  biMAP_meta_name = NULL,
  algorithm = "SNNdist",
  message = TRUE,
  ...
)

\S4method{biMAP}{cacomp,caclust}(
  obj,
  caclust_obj,
  k_umap,
  cacomp_meta_name = "caobj",
  caclust_meta_name = "caclust",
  biMAP_meta_name = NULL,
  algorithm = "SNNdist",
  message = TRUE,
  ...
)

\S4method{biMAP}{SingleCellExperiment,ANY}(
  obj,
  caclust_obj = NULL,
  k_umap,
  cacomp_meta_name = "caobj",
  caclust_meta_name = "caclust",
  biMAP_meta_name = NULL,
  algorithm = "SNNdist",
  message = TRUE,
  ...
)
}
\arguments{
\item{obj}{A cacomp object or SingleCellExperiment object}

\item{caclust_obj}{the name of caclust object stored in metadata(SingleCellExperiment object)}

\item{k_umap}{Number of nearest neighbours to use from the SNN graph for
UMAP}

\item{cacomp_meta_name}{the name of cacomp object stored in metadata(SingleCellExperiment object)}

\item{caclust_meta_name}{the name of caclust object stored in metadata(SingleCellExperiment object)}

\item{biMAP_meta_name}{character. Slot name of biMAP coordinates in metadata of sce object. Default: biMAP_$algorithm.}
}
\value{
an caclust object or SingleCellExperiment objects
}
\description{
TODO

TODO

TODO
}
\details{
TODO
}
