% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{ca_biMAP}
\alias{ca_biMAP}
\alias{ca_biMAP,caclust,cacomp-method}
\alias{ca_biMAP,SingleCellExperiment,ANY-method}
\title{Compute biMAP basedon CA results}
\usage{
ca_biMAP(
  obj,
  caobj,
  k = 30,
  rand_seed = 2358,
  use_SNN = TRUE,
  features = NULL,
  ...
)

\S4method{ca_biMAP}{caclust,cacomp}(
  obj,
  caobj,
  k = 30,
  rand_seed = 2358,
  use_SNN = TRUE,
  features = NULL,
  ...
)

\S4method{ca_biMAP}{SingleCellExperiment,ANY}(
  obj,
  caobj = NULL,
  k = 30,
  rand_seed = 2358,
  use_SNN = TRUE,
  features = NULL,
  ...,
  caclust_meta_name = "caclust",
  cacomp_meta_name = "CA"
)
}
\arguments{
\item{obj}{A `caclust` object or `SingleCellExperiment` object with `caclust`
and `cacomp` objects stored in the metadata.}

\item{caobj}{A `cacomp` object.}

\item{k}{integer. Number of nearest neighbours to use to compute UMAP.}

\item{rand_seed}{integer. Random seed for UMAP.}

\item{use_SNN}{TRUE/FALSE. This parameter only works when method == 'ca'.
If TRUE, it will calculate bimap embedding of genes and cells contained in the SNN graph together with the given 'fearures' if any.
If FALSE, only calculate bimap embedding of cells contained in the SNN graph together with the given 'fearures' if any.}

\item{features}{character/vector of gene names. This parameter only works when method == 'ca'.
'ca' method allows visualization of genes which are absent from SNN graph, so the 'features' could be genes detected by our biclustering algorithm,
and genes which goes through CA analysis. Setting use_SNN = FALSE allows users to have a visualization of the feature genes user defined/interested.}

\item{...}{Further arguments}

\item{caclust_meta_name}{the name of caclust object stored in metadata(obj)}

\item{cacomp_meta_name}{the name of cacomp object stored in
metadata(obj)}
}
\value{
an caclust object or SingleCellExperiment objects
}
\description{
The function takes either a `caclust` and a `cacomp` object as input and
computes the biMAP embedings for cells and genes on the basis of the singular
vectors of CA.
}
\details{
The biMAP embeddings are computed on the basis of the singular vectors
from Correspondence Analysis.
If a `SingleCellExperiment` object with `caclust` and `cacomp`
objects stored is provided the argument `caobj` is not required.
}
