% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinet-package.R
\docType{package}
\name{cabinet-package}
\alias{cabinet}
\alias{cabinet-package}
\title{cabinet: A package to show how to create packages}
\description{
A package created for a Cabinet Office Coffee & Coding session that doesn't do anything useful in particular.
}
\author{
\strong{Maintainer}: Matt Dray \email{matthew.dray@digital.cabinet-office.gov.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
