% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinet-package.R
\docType{package}
\name{cabinet-package}
\alias{cabinet}
\alias{cabinet-package}
\title{cabinet: A Package To Show How To Create Packages}
\description{
A package created for a Cabinet Office Coffee & Coding
    session that doesn't do anything useful in particular.
}
\author{
\strong{Maintainer}: Matt Dray \email{matthew.dray@digital.cabinet-office.gov.uk}

}
\keyword{internal}
