% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinet-package.R
\docType{package}
\name{cabinet-package}
\alias{cabinet}
\alias{cabinet-package}
\title{cabinet: Identify Cats in the UK Cabinet Office}
\description{
A package created for a Cabinet Office Coffee & Coding
    session to demonstrate the {usethis} and {devtools} packages
    for creating your own package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/cabinet}
  \item \url{https://matt-dray.github.io/cabinet/}
  \item Report bugs at \url{https://github.com/matt-dray/cabinet/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
