% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAbiNet-package.R
\docType{package}
\name{CAbiNet-package}
\alias{CAbiNet}
\alias{CAbiNet-package}
\title{CAbiNet: Biclustering and visualization of scRNAseq data}
\description{
CAbiNet allows fast and robust biclustering of single-cell RNA-seq through Correspondence Analysis. Clustering results of cells and genes can be conviently displayed in a biMAP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vingronlab.github.io/CAbiNet/}
}

}
\author{
\strong{Maintainer}: Clemens Kohl \email{kohl@molgen.mpg.de}

Authors:
\itemize{
  \item Yan Zhao \email{zhao@molgen.mpg.de}
  \item Martin Vingron \email{vingron@molgen.mpg.de}
}

}
\keyword{internal}
