% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felines.R
\name{cabinet_cat}
\alias{cabinet_cat}
\title{Check for a cabinet office cat}
\usage{
cabinet_cat(name)
}
\arguments{
\item{name}{A cat name provided as a character string}
}
\description{
Check whether a name belongs to one of the cats from the government's cabinet office
}
\examples{
cabinet_cat("Evie")
}
