% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{caclust-class}
\alias{caclust-class}
\title{An S4 class that contains all elements needed for CA.}
\description{
Class to store biclustering clustering results.
}
\section{Slots}{

\describe{
\item{\code{cell_clusters}}{factors. The assigned cell clusters with cell names in 
the names attribute.}

\item{\code{gene_clusters}}{factors. The assigned gene clusters with gene names in 
the names attribute.}

\item{\code{SNN}}{sparse shared nearest neighbours matrix. Values indicate the
jaccard similarity.}

\item{\code{parameters}}{List of used parameters and function name with which results
were generated.}

\item{\code{eigen}}{matrix, Slot for storing eigenvectors from spectral clustering}

\item{\code{cell_prob}}{matrix. Matrix that stores the probabilities that a cell belongs
to a cluster. Only filled when running spectral clustering with GMM.}

\item{\code{gene_prob}}{matrix. Matrix that stores the probabilities that a gene belongs
to a cluster. Only filled when running spectral clustering with GMM.}

\item{\code{cell_idxs}}{integer. Indices of the cells in the SNN adjacency matrix.}

\item{\code{gene_idxs}}{integer. Indices of the genes in the SNN adjacency matrix.}
}}

