\name{CAclust-package}
\alias{CAclust-package}
\alias{CAclust}
\docType{package}
\title{
\packageTitle{CAclust}
}
\description{
\packageDescription{CAclust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CAclust}
\packageIndices{CAclust}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CAclust}

Maintainer: \packageMaintainer{CAclust}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
