% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_overlap_deprecated}
\alias{calc_overlap_deprecated}
\title{WIP replacement for `determine_overlap` function}
\usage{
calc_overlap_deprecated(cc_adj, cg_adj)
}
\description{
WIP replacement for `determine_overlap` function
}
