% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{check_caobj_sce}
\alias{check_caobj_sce}
\title{check if cacomp object is already added to SingleCellExperiment object}
\usage{
check_caobj_sce(sce, cacomp_meta_name = "caobj")
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\description{
check if cacomp object is already added to SingleCellExperiment object
}
