% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ComputeSNNasym}
\alias{ComputeSNNasym}
\title{Calculate SNN graph from an adjacency matrix.}
\usage{
ComputeSNNasym(adj, mode = "all", prune = 1/15, return_dense = FALSE)
}
\arguments{
\item{adj}{A sparse adjacency matrix of type "dgCMatrix".}

\item{mode}{Whethter to use only incoming edges ("in"),
outgoing edges ("out") or both ("all").}

\item{prune}{Cutoff between 0 and 1. Nodes that have a Jaccard similarity
between them less than `prune` will be set to 0.}

\item{return_dense}{Whether to return a dense matrix or not. Overrides the
automatic decision if set to TRUE.}
}
\value{
Either a sparse matrix or a dense matrix representing
 the SNN similarity.
}
\description{
This function decides based on the size of the adjacency matrix
 whether to use a dense or sparse matrix for the results and calculates the
SNN similarity between nodes.
}
