% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeSNNasym}
\alias{ComputeSNNasym}
\title{Calculates SNN from adjacency matrix with uneven number of neighbours per
row.}
\usage{
ComputeSNNasym(SNN, prune)
}
\arguments{
\item{SNN}{A sparse matrix (adjacency matrix)}

\item{prune}{numeric. Below which Jaccard similarity edges should be 
removed.}
}
\description{
Calculates SNN from adjacency matrix with uneven number of neighbours per
row.
}
