% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{contour_plot}
\alias{contour_plot}
\title{internal helper to plot contour biMAP}
\usage{
contour_plot(
  umap_cells,
  umap_genes,
  color_by,
  color_genes,
  colors,
  interact_genes,
  gene_size,
  gene_alpha,
  contour_n
)
}
\arguments{
\item{umap_cells}{biMAP coordinates data.frame subsetted to cells.}

\item{umap_genes}{biMAP coordinates data.frame subsetted to genes.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type
(cell or gene) while "cluster" colors by the assigned cluster. If meta_df is
provided a column from the data.frame can be specified.}

\item{color_genes}{logical. If TRUE colors genes by color_by.}

\item{colors}{Vector of hexadecimal color codes of the same length as the
categories defined by color_by or longer. If NULL ignored.}

\item{interact_genes}{character string with text and labels for genes in
interactive plots.}

\item{gene_size}{integer. Size of the genes in the plot.}

\item{gene_alpha}{numeric. Alpha value of genes (between 0 and 1).}

\item{contour_n}{integer. Number of contour lines to plot.}
}
\value{
contour plot as ggplot object.
}
\description{
internal helper to plot contour biMAP
}
