% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{create_bigraph}
\alias{create_bigraph}
\title{Combine kNN graphs to large cell-gene adjecency matrix}
\usage{
create_bigraph(
  cell_dists,
  gene_dists,
  cell_gene_assr,
  gene_cell_assr,
  k_c,
  k_g,
  k_cg,
  k_gc,
  loops = FALSE,
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE
)
}
\arguments{
\item{cell_dists}{cell-cell euclidean distances}

\item{gene_dists}{gene-gene euclidean distances}

\item{cell_gene_assr}{cell-gene association ratios}

\item{gene_cell_assr}{gene-cell association ratios}

\item{k_c}{k for cell-cell kNN}

\item{k_g}{k for gene-gene kNN}

\item{k_cg}{k for cell-gene kNN}

\item{k_gc}{k for gene-cell kNN}

\item{loops}{TRUE/FALSE. If TRUE self-loops are allowed, otherwise not.}

\item{select_genes}{TRUE/FALSE. Should genes be selected by whether they have
an edge in the cell-gene kNN graph?}

\item{prune_overlap}{TRUE/FALSE. If TRUE edges to genes that share less
than `overlap` of genes with the nearest neighbours of the cell are removed.
Pruning is only performed if select_genes = TRUE.}

\item{overlap}{Numeric between 0 and 1. Overlap cutoff if
prune_overlap = TRUE.}

\item{calc_gene_cell_kNN}{TRUE/FALSE. If TRUE a cell-gene graph is calculated
by choosing the `k_gc` nearest cells for each gene. If FALSE the cell-gene
graph is transposed.}
}
\value{
"Bigraph" of type `dgCMatrix`. The combined adjacency matrix consists of the
cell-cell graph, gene-gene graph and cell-gene/gene-cell graph.
}
\description{
TODO
}
