% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{create_SNN}
\alias{create_SNN}
\title{Create SNN graph from caobj}
\usage{
create_SNN(
  caobj,
  distances,
  k_c,
  k_g,
  k_cg,
  k_gc,
  SNN_prune = 1/15,
  loops = FALSE,
  mode = "out",
  select_genes = TRUE,
  prune_overlap = TRUE,
  overlap = 0.2,
  calc_gene_cell_kNN = FALSE
)
}
\arguments{
\item{caobj}{A cacomp object with standard and principal coordinates
calculated.}

\item{distances}{A list containing the cell-cell, gene-gene, cell-gene and
gene-cell distances/association ratios. Must be named list as follows:
\itemize{
\item "cc": cell-cell euclidean distances
\item "gg": gene-gene euclidean distances
\item "cg": cell-gene association ratio
\item "gc": gene-cell association ratio
}}

\item{k_c}{k for cell-cell kNN}

\item{k_g}{k for gene-gene kNN}

\item{k_cg}{k for cell-gene kNN}

\item{k_gc}{k for gene-cell kNN}

\item{SNN_prune}{numeric. Value between 0-1. Sets cutoff of acceptable jaccard
similarity scores for neighborhood overlap of vertices in SNN. Edges with values
less than this will be set as 0. The default value is 1/15.}

\item{loops}{TRUE/FALSE. If TRUE self-loops are allowed, otherwise not.}

\item{mode}{The type of neighboring vertices to use for calculating similarity
scores(Jaccard Index). Three options: "out", "in" and "all":
\itemize{
\item "out": Selecting neighbouring vertices by out-going edges;
\item "in": Selecting neighbouring vertices by in-coming edges;
\item "all": Selecting neigbouring vertices by both in-coming and out-going edges.
}}

\item{select_genes}{TRUE/FALSE. Should genes be selected by wether they have
an edge in the cell-gene kNN graph?}

\item{prune_overlap}{TRUE/FALSE. If TRUE edges to genes that share less
than \code{overlap} of genes with the nearest neighbours of the cell are removed.}

\item{overlap}{Numeric between 0 and 1. Overlap cutoff if
prune_overlap = TRUE.}

\item{calc_gene_cell_kNN}{TRUE/FALSE. If TRUE a cell-gene graph is calculated
by choosing the `k_gc` nearest cells for each gene. If FALSE the cell-gene
graph is transposed.}

\item{calc_cell_gene_kNN}{TRUE/FALSE. If TRUE a cell-gene graph is calculated
by choosing the \code{k_gc} nearest cells for each gene. If FALSE the cell-gene
graph is transposed.}
}
\value{
A sparse adjacency Matrix of type "dgCMatrix". The values in the matrix
are the Jaccard similarity between nodes in the graph. The range between 0
and 1, with 0 meaning that no edges are shared between nodes, wheras 1 means
all edges are shared between nodes.
}
\description{
TODO
}
