% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{feature_biMAP}
\alias{feature_biMAP}
\title{plot biMAP with gene expression}
\usage{
feature_biMAP(
  sce,
  caclust,
  feature,
  color_cells_by = "expression",
  assay = "logcounts",
  gene_alpha = 0.5,
  cell_alpha = 0.8,
  gene_size = 1,
  cell_size = 1,
  label_size = 1
)
}
\arguments{
\item{sce}{SinleCellExperiment object}

\item{caclust}{caclust object}

\item{feature}{character. Name of gene to visualize}

\item{color_cells_by}{character. Default: expression}

\item{assay}{character. Name of assay in SingleCellExperiment used for visualization.}

\item{gene_alpha}{numeric. The transparent value of gene points.}

\item{cell_alpha}{numeric. The transparent value of cell points.}

\item{gene_size}{numeric. The size of gene points.}

\item{cell_size}{numeric. The size of cell points.}

\item{label_size}{numeric. Size of the text label.}
}
\description{
plot biMAP with gene expression
}
