% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{feature_biMAP}
\alias{feature_biMAP}
\title{plot biMAP with gene expression}
\usage{
feature_biMAP(
  sce,
  caclust,
  feature,
  color_cells_by = "expression",
  assay = "logcounts",
  label_size = 1
)
}
\arguments{
\item{sce}{SinleCellExperiment object}

\item{caclust}{caclust object}

\item{feature}{character. Name of gene to visualize}

\item{color_cells_by}{character. Default: expression}

\item{assay}{character. Name of assay in SingleCellExperiment used for visualization.}

\item{label_size}{numeric. Size of the text label.}
}
\description{
plot biMAP with gene expression
}
