% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{feature_biMAP}
\alias{feature_biMAP}
\title{plot biMAP with gene expression}
\usage{
feature_biMAP(
  umap_coords,
  sce,
  feature = NULL,
  color_cells_by = "expression",
  assay = "logcounts"
)
}
\description{
plot biMAP with gene expression
}
