% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hex_plot}
\alias{hex_plot}
\title{internal helper to plot hex biMAP}
\usage{
hex_plot(
  umap_cells,
  umap_genes,
  color_by,
  colors,
  color_genes,
  hex_n,
  min_bin,
  show_density,
  cell_alpha,
  gene_size,
  gene_alpha,
  interact_genes
)
}
\arguments{
\item{umap_cells}{biMAP coordinates data.frame subsetted to cells.}

\item{umap_genes}{biMAP coordinates data.frame subsetted to genes.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type 
(cell or gene) while "cluster" colors by the assigned cluster. If meta_df is
provided a column from the data.frame can be specified.}

\item{colors}{Vector of hexadecimal color codes of the same length as the
categories defined by color_by or longer. If NULL ignored.}

\item{color_genes}{logical. If TRUE colors genes by color_by.}

\item{hex_n}{integer. Number of hexagonal bins that partition the x-axis.}

\item{min_bin}{integer. Only bins with more than `min_bin` cells are plotted.}

\item{show_density}{logical. Should the alpha value for plot type "hex" be
adjusted by the number of cells in the bin?}

\item{cell_alpha}{numeric. Alpha value of cells (between 0 and 1).}

\item{gene_size}{integer. Size of the genes in the plot.}

\item{gene_alpha}{numeric. Alpha value of genes (between 0 and 1).}

\item{interact_genes}{character string with text and labels for genes in 
interactive plots.}
}
\value{
ggplot object
}
\description{
internal helper to plot hex biMAP
}
