% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_biMAP}
\alias{plot_biMAP}
\alias{plot_biMAP,caclust-method}
\alias{plot_biMAP,SingleCellExperiment-method}
\title{Plotting biMAP embedding}
\usage{
plot_biMAP(
  obj,
  caclust_meta_name = "caclust",
  meta_df = NULL,
  color_by = "cluster",
  type = "scatter",
  cell_size = 1,
  gene_size = 3,
  hex_n = 40,
  min_bin = 2,
  contour_n = 5,
  cell_alpha = 0.5,
  gene_alpha = 1,
  show_density = FALSE,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  labels_per_group = 1,
  label_marker_genes = FALSE,
  ...
)

\S4method{plot_biMAP}{caclust}(
  obj,
  caclust_meta_name = "caclust",
  meta_df = NULL,
  color_by = "cluster",
  type = "scatter",
  cell_size = 1,
  gene_size = 3,
  hex_n = 40,
  min_bin = 2,
  contour_n = 5,
  cell_alpha = 0.5,
  gene_alpha = 1,
  show_density = FALSE,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  labels_per_group = 1,
  label_marker_genes = FALSE,
  ...
)

\S4method{plot_biMAP}{SingleCellExperiment}(
  obj,
  caclust_meta_name = "caclust",
  meta_df = NULL,
  color_by = "cluster",
  type = "scatter",
  cell_size = 1,
  gene_size = 3,
  hex_n = 40,
  min_bin = 2,
  contour_n = 5,
  cell_alpha = 0.5,
  gene_alpha = 1,
  show_density = FALSE,
  color_genes = FALSE,
  label_groups = TRUE,
  group_label_size = 4,
  labels_per_group = 1,
  label_marker_genes = FALSE,
  ...
)
}
\arguments{
\item{obj}{SingleCellExperiment object}

\item{caclust_meta_name}{character. Slot name of caclust object stored in meatadata(obj)}

\item{meta_df}{data.frame.}

\item{color_by}{character which can be chosen from 'type', 'cluster',column in the input meta_df and columns in colData of the obj (if meta_df == NULL)}
}
\value{
A caclust object or SingleCellExperiment object.
}
\description{
TODO
}
\details{
TODO
}
