% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_feature_biMAP}
\alias{plot_feature_biMAP}
\alias{plot_feature_biMAP,SingleCellExperiment,caclust-method}
\alias{plot_feature_biMAP,SingleCellExperiment,ANY-method}
\title{BiMAP visualization of feature expression.}
\usage{
plot_feature_biMAP(
  sce,
  caclust = NULL,
  feature,
  color_cells_by = "expression",
  assay = "logcounts",
  ...
)

\S4method{plot_feature_biMAP}{SingleCellExperiment,caclust}(
  sce,
  caclust = NULL,
  feature,
  color_cells_by = "expression",
  assay = "logcounts",
  ...
)

\S4method{plot_feature_biMAP}{SingleCellExperiment,ANY}(
  sce,
  caclust = NULL,
  feature,
  color_cells_by = "expression",
  assay = "logcounts",
  ...,
  caclust_meta_name = "caclust"
)
}
\arguments{
\item{sce}{SinleCellExperiment object}

\item{caclust}{caclust object with bimap slot or NULL}

\item{feature}{character. Name of gene to visualize}

\item{color_cells_by}{character. Default: expression}

\item{assay}{character. Name of assay in SingleCellExperiment used for visualization.}

\item{...}{Further arguments}

\item{caclust_meta_name}{Name under which the caclust object is stored in
the metadata of the SingleCellExperiment object.}
}
\value{
ggplot object
}
\description{
BiMAP visualization of feature expression.
}
