% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rm_monoclusters}
\alias{rm_monoclusters}
\alias{rm_monoclusters,caclust-method}
\alias{rm_monoclusters,Biclust-method}
\alias{rm_monoclusters,SingleCellExperiment-method}
\title{Remove clusters only consisting of cells/genes}
\usage{
rm_monoclusters(obj, ...)

\S4method{rm_monoclusters}{caclust}(obj)

\S4method{rm_monoclusters}{Biclust}(obj)

\S4method{rm_monoclusters}{SingleCellExperiment}(obj, subset_sce = FALSE)
}
\arguments{
\item{obj}{biclustering results from caclust or biclust or a SingleCellExperiment object.}

\item{subset_sce}{logical. Subset the SCE object to the cells which have co-clusters genes if is TRUE.}

\item{sce}{SingleCellExperiment object with caclust object in the metadata}
}
\value{
caclust/biclust object with monoclusters removed.
}
\description{
Takes an object of class biclust and removes all clusters that only consist
of cells or genes.
}
