% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biMAP.R
\name{run_biMAP}
\alias{run_biMAP}
\title{Plots UMAP depicting both cells and genes.}
\usage{
run_biMAP(caobj, caclust_obj, k_umap, rand_seed = 2358, algorithm = "SNNdist")
}
\arguments{
\item{caobj}{A cacomp object with principal and standard coordinates 
calculated.}

\item{caclust_obj}{results from biclustering of class "caclust"}

\item{k_umap}{Number of nearest neighbours to use from the SNN graph for
UMAP}

\item{rand_seed}{Random seed for UMAP.}
}
\value{
caclust object with biMAP coordinates stored in the `bimap` slot.
}
\description{
TODO
}
