% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_leiden}
\alias{run_leiden}
\title{Leiden clustering on bigraph}
\usage{
run_leiden(caclust, resolution = 1, n.int = 10, rand_seed = 2358, dense = TRUE)
}
\arguments{
\item{resolution}{resolution for leiden algorithm.}

\item{n.int}{Number of iterations for leiden algorithm.}

\item{rand_seed}{Random seed.}

\item{SNN}{dense or sparse matrix. A SNN graph.}
}
\value{
vector of type `factor.` Assigned clusters of cells and genes. 
The names of cells and genes are saved in the names of the 
vector.
}
\description{
TODO
}
