% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_leiden}
\alias{run_leiden}
\title{Leiden clustering on bigraph}
\usage{
run_leiden(
  caclust,
  resolution = 1,
  n.int = 10,
  rand_seed = 2358,
  cast_to_dense = TRUE,
  leiden_pack = "igraph"
)
}
\arguments{
\item{caclust}{caclust object with SNN calculated.}

\item{resolution}{float number. Resolution for leiden algorithm.}

\item{n.int}{Integer. Number of iterations for leiden algorithm.}

\item{rand_seed}{integer. Random seed.}

\item{cast_to_dense}{logical. Should the SNN-graph be converted to a dense}

\item{leiden_pack}{character. Optional values are 'igraph'(default) and 'leiden', the package used for leiden clustering.
matrix before running leiden clustering?
Casting to dense speeds up the leiden algorithm.}
}
\value{
Object of type "caclust" with cell and gene clusters saved.
}
\description{
This function takes a caclust object with precomputed SNN-graph and
clusters cells and genes simultaneously.
}
\seealso{
Other biclustering: 
\code{\link{caclust}},
\code{\link{make_SNN}()},
\code{\link{run_spectral}()}
}
\concept{biclustering}
