% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{internal helper to plot scatter biMAP}
\usage{
scatter_plot(
  umap_cells,
  umap_genes,
  color_by,
  colors,
  color_genes,
  cell_size,
  cell_alpha,
  gene_size,
  gene_alpha,
  interact_genes,
  interact_cells
)
}
\arguments{
\item{umap_cells}{biMAP coordinates data.frame subsetted to cells.}

\item{umap_genes}{biMAP coordinates data.frame subsetted to genes.}

\item{color_by}{Either "type" or "cluster". "type" colors by the type
(cell or gene) while "cluster" colors by the assigned cluster. If meta_df is
provided a column from the data.frame can be specified.}

\item{colors}{named vector of of length equal to the group size.}

\item{color_genes}{logical. If TRUE colors genes by color_by.}

\item{cell_size}{integer. Size of the cells in the plot.}

\item{cell_alpha}{numeric. Alpha value of cells (between 0 and 1).}

\item{gene_size}{integer. Size of the genes in the plot.}

\item{gene_alpha}{numeric. Alpha value of genes (between 0 and 1).}

\item{interact_genes}{character string with text and labels for genes in
interactive plots.}

\item{interact_cells}{character string with text and labels for cells in
interactive plot.}
}
\value{
ggplot object
}
\description{
internal helper to plot scatter biMAP
}
