% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{SKMeans}
\alias{SKMeans}
\title{An integration of several runs of skmens with different random seeds}
\usage{
SKMeans(
  x,
  k,
  method = NULL,
  m = 1,
  weights = 1,
  control = list(),
  num.seeds = 10
)
}
\arguments{
\item{method}{see skmeans method}

\item{m}{see skmeans 'm'}

\item{num.seeds}{number of trials with random seeds}
}
\value{
The optimal skmeans clustering result
}
\description{
This function will select the optimal clustering result from several 
skmeans::skmeans runs
with different random seeds, The clustering result with smallest 
within-cluster-sum
of squared distances will be selected.
}
